Separated InputHandler.java into subclasses for improved readibility
