+++ src/main/java/jterm/io/ArrowKeys.java
@@ -1,0 +1,5 @@
+package jterm.io;
+
+public enum ArrowKeys {
+    UP, DOWN, LEFT, RIGHT, NONE, MOD // for info on MOD use, please see ProcessArrowKey()
+}
