+++ src/main/java/jterm/io/Keys.java
@@ -1,0 +1,5 @@
+package jterm.io;
+
+public enum Keys {
+    TAB, BCKSP, NWLN, NONE
+}
